+++ source/extensions/omni.isaac.manipulators/omni/isaac/manipulators/manipulators/__init__.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2021-2024, NVIDIA CORPORATION. All rights reserved.
+# Copyright (c) 2018-2024, NVIDIA CORPORATION. All rights reserved.
 #
 # NVIDIA CORPORATION and its licensors retain all intellectual property
 # and proprietary rights in and to this software, related documentation
@@ -6,5 +6,6 @@
 # distribution of this software and related documentation without an express
 # license agreement from NVIDIA CORPORATION is strictly prohibited.
 #
-from .extension import Extension
-from .manipulators.single_manipulator import SingleManipulator
+
+
+from .single_manipulator import SingleManipulator