+++ source/extensions/omni.isaac.ui/omni/isaac/ui/widgets.py
@@ -8,12 +8,15 @@
 #
 
 
+from collections import namedtuple
 from typing import Callable, Optional
 
 import carb
 import omni
 import omni.ui as ui
 from omni.isaac.ui.ui_utils import BUTTON_WIDTH, LABEL_WIDTH, get_style
+from omni.kit.property.usd.relationship import RelationshipTargetPicker
+from omni.kit.window.popup_dialog.dialog import get_field_value
 
 
 class DynamicComboBoxItem(ui.AbstractItem):
@@ -44,3 +47,84 @@ class DynamicComboBoxModel(ui.AbstractItemModel):
         self._current_index = item
         self._item_changed(None)
         self._current_index.add_value_changed_fn(lambda a: self._item_changed(None))
+
+
+class SelectPrimWidget:
+    """
+    modeled after FormWidget (from omni.kit.window.popup_dialog.form_dialog) to add a widget that opens relationship selector
+    """
+
+    def __init__(self, label: str = None, default: str = None):
+        self._label = label
+        self._default_path = default
+        self._build_ui()
+
+    def _build_ui(self):
+        with ui.HStack(height=0):
+            ui.Label(
+                self._label, width=ui.Percent(29), style_type_name_override="Field.Label", word_wrap=True, name="prefix"
+            )
+            ui.Spacer(width=ui.Percent(1))
+            path_model = ui.SimpleStringModel()
+            path_model.set_value(self._default_path)
+            self._prim = ui.StringField(path_model, width=ui.Percent(50))
+            ui.Spacer(width=ui.Percent(1))
+            ui.Button("Add", width=ui.Percent(19), clicked_fn=self._on_select_prim)
+
+    def _on_select_prim(self):
+        stage = omni.usd.get_context().get_stage()
+        additional_widget_kwargs = {"target_name": "Prim"}
+        self.stage_picker = RelationshipTargetPicker(
+            stage,
+            [],
+            None,
+            additional_widget_kwargs,
+        )
+        self.stage_picker.show(1, on_targets_selected=self._on_target_selected)
+
+    def _on_target_selected(self, paths):
+        self._prim.model.set_value(paths[0])
+
+    def get_value(self):
+        return self._prim.model.get_value_as_string()
+
+    def destroy(self):
+        self._prim = None
+
+
+class ParamWidget:
+    """
+        modified FormWidget (from omni.kit.window.popup_dialog.form_dialog) to better format for parameter collection use
+
+    Note:
+        ParamWidget.FieldDef:
+            A namedtuple of (name, label, type, default value) for describing the input field,
+            e.g. FormDialog.FieldDef("name", "Name:  ", omni.ui.StringField, "Bob").
+
+    """
+
+    FieldDef = namedtuple("FormDialogFieldDef", "name label type default focused", defaults=[False])
+
+    def __init__(self, field_def: FieldDef = ()):
+        self._field = None
+        self._build_ui(field_def)
+
+    def _build_ui(self, field_def):
+        with ui.HStack(height=0):
+            ui.Label(
+                field_def.label,
+                width=ui.Percent(29),
+                style_type_name_override="Field.Label",
+                word_wrap=True,
+                name="prefix",
+            )
+            ui.Spacer(width=ui.Percent(1))
+            self._field = field_def.type(width=ui.Percent(70))
+            if "set_value" in dir(self._field.model):
+                self._field.model.set_value(field_def.default)
+
+    def get_value(self):
+        return get_field_value(self._field)
+
+    def destroy(self):
+        self._field = None